Memoji injection - Gioco-1

+++ Assets/Scripts/MedicalReport.cs
@@ -32,7 +32,7 @@ namespace Application {
 
       pathology = new Pathology();
 
-      animojiPath = "Sprites/" + name + "_bad";
+      animojiPath = "Sprites/" + (int) name + "_bad";
     }
 
     public override string ToString(){
